#226 [fix] 인기 프로젝트 목록 조회 기능을 search에서 read로 변경한다.
