+++ src/main/java/com/dataracy/modules/project/adapter/web/api/read/ProjectReadApi.java
@@ -3,6 +3,7 @@ package com.dataracy.modules.project.adapter.web.api.read;
 import com.dataracy.modules.common.dto.response.SuccessResponse;
 import com.dataracy.modules.project.adapter.web.response.read.ConnectedProjectWebResponse;
 import com.dataracy.modules.project.adapter.web.response.read.ContinuedProjectWebResponse;
+import com.dataracy.modules.project.adapter.web.response.read.PopularProjectWebResponse;
 import com.dataracy.modules.project.adapter.web.response.read.ProjectDetailWebResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -22,6 +23,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+
+import java.util.List;
 
 @Tag(name = "Project", description = "프로젝트 관련 API")
 @RequestMapping("/api/v1/projects")
@@ -101,5 +104,27 @@ public interface ProjectReadApi {
 
             @PageableDefault(size = 3, page = 0)
             Pageable pageable
+    );
+
+    /**
+     * 좋아요, 댓글, 조회수를 기준으로 인기 프로젝트 목록을 조회한다.
+     *
+     * @param size 반환할 최대 프로젝트 개수 (1 이상)
+     * @return 인기 프로젝트 목록이 포함된 성공 응답 객체
+     */
+    @Operation(
+            summary = "인기있는 프로젝트 리스트를 조회한다.",
+            description = "좋아요, 댓글, 조회수를 기준으로 프로젝트 리스트를 조회한다."
+    )
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "인기있는 프로젝트 리스트 조회에 성공했습니다.",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = SuccessResponse.class)))
+    })
+    @GetMapping("/popular")
+    ResponseEntity<SuccessResponse<List<PopularProjectWebResponse>>> searchPopularProjects(
+            @RequestParam(name = "size")
+            @Min(1)
+            int size
     );
 }
